==> 00-county-meta.Rmd <==
---
title: County metadata
output: github_document
---

The purpose of this document is to write out the 2019 estimate for the population of Iowa counties, along with other metadata. 

The source of this data is the [Iowa Community Indicators Program](https://www.icip.iastate.edu/tables/population/counties-estimates).

```{r packages}
library("fs")
library("magrittr")
library("iowa.covid")
library("dplyr")
library("readr")
library("ggplot2")
library("stringr")
```

```{r dir}
dir_create("data")

dir_target <- path("data", "meta")
dir_create(dir_target)
```

In late April, Gov. Reynolds announced that she would partially re-open 77 counties effective May 1. The remaining 22 counties were similarly re-opened May 15. These are the 22 "later" counties:

```{r}
county_closed <- 
  c(
    "Allamakee", 
    "Benton", 
    "Black Hawk", 
    "Bremer", 
    "Dallas", 
    "Des Moines", 
    "Dubuque", 
    "Fayette", 
    "Henry", 
    "Iowa", 
    "Jasper", 
    "Johnson", 
    "Linn", 
    "Louisa", 
    "Marshall", 
    "Muscatine", 
    "Polk", 
    "Poweshiek", 
    "Scott", 
    "Tama", 
    "Washington", 
    "Woodbury"
  )
```

We also want to compute the approximate centers of the counties.

```{r}
map_iowa_county <-
  map_data("county") %>%
  filter(region == "iowa") %>%
  group_by(county = subregion) %>%
  summarise(
    lat = (max(lat) + min(lat)) / 2,
    lon =  (max(long) + min(long)) / 2
  ) %>%
  mutate(
    abbreviation = str_to_upper(county) %>% str_sub(1, 3),
    county = str_to_title(county),
    county = ifelse(county == "Obrien", "O'Brien", county)
  ) %>%
  glimpse()
```

```{r}
iowa_county_meta <-
  iowa_county_population %>%
  select(fips, county, population, population_group) %>%
  mutate(
    opening = 
      ifelse(county %in% county_closed, "2020-05-15", "2020-05-01") %>%
      as.Date()
  ) %>%
  left_join(
    map_iowa_county, 
    by = "county"
  ) %>%
  select(fips, county, abbreviation, lon, lat, everything()) %>%
  glimpse()
```


```{r}
write_csv(
  iowa_county_meta, 
  path(dir_target, "iowa_county_meta.csv")
)
```
